Fix import cycle by extracting time sync to separate package
